Extract use-single-result-page-meta composable
